// src/app/feature/profiles-settings/model/profile-settings-model.ts
export interface ProfileSettingsModel {
  id?: string;
  code?: string;
  name?: string;
  description?: string;
  price?: number;
  quantity?: number;
  inventoryStatus?: InventoryStatus;
  category?: string;
  image?: string;
  rating?: number;
}
interface InventoryStatus {
  label: string;
  value: string;
}